chore: mega code cleanup
- update AttributeManagerUtil to generate smell free code
- other minor lanaguage specific changes everywhere

+++ soap/src/main/java/com/zimbra/soap/admin/type/IntegerValueAttrib.java
@@ -26,7 +26,7 @@ public class IntegerValueAttrib {
      */
     @SuppressWarnings("unused")
     private IntegerValueAttrib() {
-        this((Integer) null);
+        this(null);
     }
 
     public IntegerValueAttrib(Integer value) {
